rag-backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
python-dotenv>=1.0.0
python-multipart>=0.0.9

# Embeddings
sentence-transformers>=3.0.0
torch>=2.0.0
scikit-learn>=1.4.0
numpy>=1.26.0

# Text extraction
pymupdf>=1.24.0
python-docx>=1.1.0

# LLMs
google-generativeai>=0.8.0
openai>=1.30.0

# Supabase + pgvector
supabase>=2.4.0

# Google Drive
google-auth-oauthlib>=1.2.0
google-api-python-client>=2.130.0

# NLP (for semantic chunking)
nltk>=3.8.0
